src/requirements.txt
blessed==1.17.6
inquirer==2.7.0
Pillow==8.2.0
prompt-tool-kit==1.0.14
pycryptodome==3.10.1
Pygments==2.9.0
python-editor==1.0.4
readchar==2.0.1
regex==2021.4.4
six==1.16.0
termcolor==1.1.0
wcwidth==0.2.5
